Convert Marker List to a Weighted Matrix

This function converts a list of marker genes for different cell types or clusters into a matrix, where rows represent cell types and columns represent marker genes. The matrix contains the frequency of each marker gene across the different cell types.

markerList2mat(markerList)

Arguments

markerList

A list where each element contains marker genes for different cell types or clusters. Each element of the list is a named list, where the names are the cell types and the values are the marker genes for that cell type. Generally, it is a list of the output of function marker.select.

Value

A matrix with rows representing cell types and columns representing unique marker genes. The values in the matrix represent the frequency of each gene as a marker in the corresponding cell type.

See also

Examples

data(toydata)

markers <- toydata$markers

marker.freq <- markerList2mat(list(markers))
print(marker.freq)
#>                   CCDC80 CRISPLD2 DPT FBLN1 LUM MMP2 PDGFRA PTGDS AGR3 ESR1
#> CAFs                   1        1   1     1   1    1      1     1    0    0
#> Cancer Epithelial      0        0   0     0   0    0      0     0    1    1
#> Endothelial            0        0   0     0   0    0      0     0    0    0
#> Myeloid                0        0   0     0   0    0      0     0    0    0
#> Normal Epithelial      0        0   0     0   0    0      0     0    0    0
#> Plasmablasts           0        0   0     0   0    0      0     0    0    0
#> PVL                    0        0   0     0   0    0      0     0    0    0
#> T-cells                0        0   0     0   0    0      0     0    0    0
#>                   FASN FOXA1 LYPD3 MLPH RHOH SCD BTNL9 EGFL7 HOXD9 IL3RA KDR
#> CAFs                 0     0     0    0    0   0     0     0     0     0   0
#> Cancer Epithelial    1     1     1    1    1   1     0     0     0     0   0
#> Endothelial          0     0     0    0    0   0     1     1     1     1   1
#> Myeloid              0     0     0    0    0   0     0     0     0     0   0
#> Normal Epithelial    0     0     0    0    0   0     0     0     0     0   0
#> Plasmablasts         0     0     0    0    0   0     0     0     0     0   0
#> PVL                  0     0     0    0    0   0     0     0     0     0   0
#> T-cells              0     0     0    0    0   0     0     0     0     0   0
#>                   SOX17 SOX18 VWF AIF1 C1QC CD86 FCER1G HAVCR2 IGSF6 ITGAX MNDA
#> CAFs                  0     0   0    0    0    0      0      0     0     0    0
#> Cancer Epithelial     0     0   0    0    0    0      0      0     0     0    0
#> Endothelial           1     1   1    0    0    0      0      0     0     0    0
#> Myeloid               0     0   0    1    1    1      1      1     1     1    1
#> Normal Epithelial     0     0   0    0    0    0      0      0     0     0    0
#> Plasmablasts          0     0   0    0    0    0      0      0     0     0    0
#> PVL                   0     0   0    0    0    0      0      0     0     0    0
#> T-cells               0     0   0    0    0    0      0      0     0     0    0
#>                   C5orf46 KRT14 KRT15 KRT16 KRT5 KRT6B MYH11 PIGR CD27 CD79A
#> CAFs                    0     0     0     0    0     0     0    0    0     0
#> Cancer Epithelial       0     0     0     0    0     0     0    0    0     0
#> Endothelial             0     0     0     0    0     0     0    0    0     0
#> Myeloid                 0     0     0     0    0     0     0    0    0     0
#> Normal Epithelial       1     1     1     1    1     1     1    1    0     0
#> Plasmablasts            0     0     0     0    0     0     0    0    1     1
#> PVL                     0     0     0     0    0     0     0    0    0     0
#> T-cells                 0     0     0     0    0     0     0    0    0     0
#>                   DERL3 ITM2C MZB1 PRDM1 SLAMF7 TNFRSF17 ANGPT2 AQP1 AVPR1A
#> CAFs                  0     0    0     0      0        0      0    0      0
#> Cancer Epithelial     0     0    0     0      0        0      0    0      0
#> Endothelial           0     0    0     0      0        0      0    0      0
#> Myeloid               0     0    0     0      0        0      0    0      0
#> Normal Epithelial     0     0    0     0      0        0      0    0      0
#> Plasmablasts          1     1    1     1      1        1      0    0      0
#> PVL                   0     0    0     0      0        0      1    1      1
#> T-cells               0     0    0     0      0        0      0    0      0
#>                   CAV1 EGFR PDGFRB TCEAL7 ZEB1 CCL5 CD247 CD3D CD3E CD3G CD69
#> CAFs                 0    0      0      0    0    0     0    0    0    0    0
#> Cancer Epithelial    0    0      0      0    0    0     0    0    0    0    0
#> Endothelial          0    0      0      0    0    0     0    0    0    0    0
#> Myeloid              0    0      0      0    0    0     0    0    0    0    0
#> Normal Epithelial    0    0      0      0    0    0     0    0    0    0    0
#> Plasmablasts         0    0      0      0    0    0     0    0    0    0    0
#> PVL                  1    1      1      1    1    0     0    0    0    0    0
#> T-cells              0    0      0      0    0    1     1    1    1    1    1
#>                   GZMA KLRB1
#> CAFs                 0     0
#> Cancer Epithelial    0     0
#> Endothelial          0     0
#> Myeloid              0     0
#> Normal Epithelial    0     0
#> Plasmablasts         0     0
#> PVL                  0     0
#> T-cells              1     1